+++ lib/src/pigeon/messages.pigeon.dart
@@ -25,20 +25,20 @@ List<Object?> wrapResponse({Object? result, PlatformException? error, bool empty
   }
   return <Object?>[error.code, error.message, error.details];
 }
-
 bool _deepEquals(Object? a, Object? b) {
   if (a is List && b is List) {
-    return a.length == b.length && a.indexed.every(((int, dynamic) item) => _deepEquals(item.$2, b[item.$1]));
+    return a.length == b.length &&
+        a.indexed
+        .every(((int, dynamic) item) => _deepEquals(item.$2, b[item.$1]));
   }
   if (a is Map && b is Map) {
-    return a.length == b.length &&
-        a.entries.every(
-          (MapEntry<Object?, Object?> entry) =>
-              (b as Map<Object?, Object?>).containsKey(entry.key) && _deepEquals(entry.value, b[entry.key]),
-        );
+    return a.length == b.length && a.entries.every((MapEntry<Object?, Object?> entry) =>
+        (b as Map<Object?, Object?>).containsKey(entry.key) &&
+        _deepEquals(entry.value, b[entry.key]));
   }
   return a == b;
 }
+
 
 enum ReclaimApiVerificationExceptionType {
   unknown,
@@ -60,7 +60,10 @@ enum ReclaimSessionStatus {
   PROOF_MANUAL_VERIFICATION_SUBMITTED,
 }
 
-enum ClaimCreationTypeApi { standalone, meChain }
+enum ClaimCreationTypeApi {
+  standalone,
+  meChain,
+}
 
 class ReclaimApiVerificationRequest {
   ReclaimApiVerificationRequest({
@@ -74,6 +77,7 @@ class ReclaimApiVerificationRequest {
     required this.parameters,
     required this.acceptAiProviders,
     this.webhookUrl,
+    this.providerVersion,
   });
 
   String appId;
@@ -96,6 +100,8 @@ class ReclaimApiVerificationRequest {
 
   String? webhookUrl;
 
+  ProviderVersionApi? providerVersion;
+
   List<Object?> _toList() {
     return <Object?>[
       appId,
@@ -108,12 +114,12 @@ class ReclaimApiVerificationRequest {
       parameters,
       acceptAiProviders,
       webhookUrl,
+      providerVersion,
     ];
   }
 
   Object encode() {
-    return _toList();
-  }
+    return _toList();  }
 
   static ReclaimApiVerificationRequest decode(Object result) {
     result as List<Object?>;
@@ -128,6 +134,7 @@ class ReclaimApiVerificationRequest {
       parameters: (result[7] as Map<Object?, Object?>?)!.cast<String, String>(),
       acceptAiProviders: result[8]! as bool,
       webhookUrl: result[9] as String?,
+      providerVersion: result[10] as ProviderVersionApi?,
     );
   }
 
@@ -145,11 +152,16 @@ class ReclaimApiVerificationRequest {
 
   @override
   // ignore: avoid_equals_and_hash_code_on_mutable_classes
-  int get hashCode => Object.hashAll(_toList());
+  int get hashCode => Object.hashAll(_toList())
+;
 }
 
 class ReclaimApiVerificationException {
-  ReclaimApiVerificationException({required this.message, required this.stackTraceAsString, required this.type});
+  ReclaimApiVerificationException({
+    required this.message,
+    required this.stackTraceAsString,
+    required this.type,
+  });
 
   String message;
 
@@ -158,12 +170,15 @@ class ReclaimApiVerificationException {
   ReclaimApiVerificationExceptionType type;
 
   List<Object?> _toList() {
-    return <Object?>[message, stackTraceAsString, type];
+    return <Object?>[
+      message,
+      stackTraceAsString,
+      type,
+    ];
   }
 
   Object encode() {
-    return _toList();
-  }
+    return _toList();  }
 
   static ReclaimApiVerificationException decode(Object result) {
     result as List<Object?>;
@@ -188,7 +203,8 @@ class ReclaimApiVerificationException {
 
   @override
   // ignore: avoid_equals_and_hash_code_on_mutable_classes
-  int get hashCode => Object.hashAll(_toList());
+  int get hashCode => Object.hashAll(_toList())
+;
 }
 
 class ReclaimApiVerificationResponse {
@@ -208,12 +224,16 @@ class ReclaimApiVerificationResponse {
   ReclaimApiVerificationException? exception;
 
   List<Object?> _toList() {
-    return <Object?>[sessionId, didSubmitManualVerification, proofs, exception];
+    return <Object?>[
+      sessionId,
+      didSubmitManualVerification,
+      proofs,
+      exception,
+    ];
   }
 
   Object encode() {
-    return _toList();
-  }
+    return _toList();  }
 
   static ReclaimApiVerificationResponse decode(Object result) {
     result as List<Object?>;
@@ -239,14 +259,15 @@ class ReclaimApiVerificationResponse {
 
   @override
   // ignore: avoid_equals_and_hash_code_on_mutable_classes
-  int get hashCode => Object.hashAll(_toList());
+  int get hashCode => Object.hashAll(_toList())
+;
 }
 
 class ClientProviderInformationOverride {
   ClientProviderInformationOverride({
     this.providerInformationUrl,
     this.providerInformationJsonString,
-    required this.canFetchProviderInformationFromHost,
+    this.canFetchProviderInformationFromHost = false,
   });
 
   String? providerInformationUrl;
@@ -256,12 +277,15 @@ class ClientProviderInformationOverride {
   bool canFetchProviderInformationFromHost;
 
   List<Object?> _toList() {
-    return <Object?>[providerInformationUrl, providerInformationJsonString, canFetchProviderInformationFromHost];
+    return <Object?>[
+      providerInformationUrl,
+      providerInformationJsonString,
+      canFetchProviderInformationFromHost,
+    ];
   }
 
   Object encode() {
-    return _toList();
-  }
+    return _toList();  }
 
   static ClientProviderInformationOverride decode(Object result) {
     result as List<Object?>;
@@ -286,7 +310,8 @@ class ClientProviderInformationOverride {
 
   @override
   // ignore: avoid_equals_and_hash_code_on_mutable_classes
-  int get hashCode => Object.hashAll(_toList());
+  int get hashCode => Object.hashAll(_toList())
+;
 }
 
 class ClientFeatureOverrides {
@@ -297,6 +322,7 @@ class ClientFeatureOverrides {
     this.sessionTimeoutForManualVerificationTrigger,
     this.attestorBrowserRpcUrl,
     this.isAIFlowEnabled,
+    this.manualReviewMessage,
   });
 
   bool? cookiePersist;
@@ -311,6 +337,8 @@ class ClientFeatureOverrides {
 
   bool? isAIFlowEnabled;
 
+  String? manualReviewMessage;
+
   List<Object?> _toList() {
     return <Object?>[
       cookiePersist,
@@ -319,12 +347,12 @@ class ClientFeatureOverrides {
       sessionTimeoutForManualVerificationTrigger,
       attestorBrowserRpcUrl,
       isAIFlowEnabled,
+      manualReviewMessage,
     ];
   }
 
   Object encode() {
-    return _toList();
-  }
+    return _toList();  }
 
   static ClientFeatureOverrides decode(Object result) {
     result as List<Object?>;
@@ -335,6 +363,7 @@ class ClientFeatureOverrides {
       sessionTimeoutForManualVerificationTrigger: result[3] as int?,
       attestorBrowserRpcUrl: result[4] as String?,
       isAIFlowEnabled: result[5] as bool?,
+      manualReviewMessage: result[6] as String?,
     );
   }
 
@@ -352,14 +381,15 @@ class ClientFeatureOverrides {
 
   @override
   // ignore: avoid_equals_and_hash_code_on_mutable_classes
-  int get hashCode => Object.hashAll(_toList());
+  int get hashCode => Object.hashAll(_toList())
+;
 }
 
 class ClientLogConsumerOverride {
   ClientLogConsumerOverride({
-    required this.enableLogHandler,
-    required this.canSdkCollectTelemetry,
-    this.canSdkPrintLogs,
+    this.enableLogHandler = true,
+    this.canSdkCollectTelemetry = true,
+    this.canSdkPrintLogs = false,
   });
 
   bool enableLogHandler;
@@ -369,12 +399,15 @@ class ClientLogConsumerOverride {
   bool? canSdkPrintLogs;
 
   List<Object?> _toList() {
-    return <Object?>[enableLogHandler, canSdkCollectTelemetry, canSdkPrintLogs];
+    return <Object?>[
+      enableLogHandler,
+      canSdkCollectTelemetry,
+      canSdkPrintLogs,
+    ];
   }
 
   Object encode() {
-    return _toList();
-  }
+    return _toList();  }
 
   static ClientLogConsumerOverride decode(Object result) {
     result as List<Object?>;
@@ -399,25 +432,31 @@ class ClientLogConsumerOverride {
 
   @override
   // ignore: avoid_equals_and_hash_code_on_mutable_classes
-  int get hashCode => Object.hashAll(_toList());
+  int get hashCode => Object.hashAll(_toList())
+;
 }
 
 class ClientReclaimSessionManagementOverride {
-  ClientReclaimSessionManagementOverride({required this.enableSdkSessionManagement});
+  ClientReclaimSessionManagementOverride({
+    this.enableSdkSessionManagement = true,
+  });
 
   bool enableSdkSessionManagement;
 
   List<Object?> _toList() {
-    return <Object?>[enableSdkSessionManagement];
+    return <Object?>[
+      enableSdkSessionManagement,
+    ];
   }
 
   Object encode() {
-    return _toList();
-  }
+    return _toList();  }
 
   static ClientReclaimSessionManagementOverride decode(Object result) {
     result as List<Object?>;
-    return ClientReclaimSessionManagementOverride(enableSdkSessionManagement: result[0]! as bool);
+    return ClientReclaimSessionManagementOverride(
+      enableSdkSessionManagement: result[0]! as bool,
+    );
   }
 
   @override
@@ -434,11 +473,16 @@ class ClientReclaimSessionManagementOverride {
 
   @override
   // ignore: avoid_equals_and_hash_code_on_mutable_classes
-  int get hashCode => Object.hashAll(_toList());
+  int get hashCode => Object.hashAll(_toList())
+;
 }
 
 class ClientReclaimAppInfoOverride {
-  ClientReclaimAppInfoOverride({required this.appName, required this.appImageUrl, required this.isRecurring});
+  ClientReclaimAppInfoOverride({
+    required this.appName,
+    required this.appImageUrl,
+    required this.isRecurring,
+  });
 
   String appName;
 
@@ -447,12 +491,15 @@ class ClientReclaimAppInfoOverride {
   bool isRecurring;
 
   List<Object?> _toList() {
-    return <Object?>[appName, appImageUrl, isRecurring];
+    return <Object?>[
+      appName,
+      appImageUrl,
+      isRecurring,
+    ];
   }
 
   Object encode() {
-    return _toList();
-  }
+    return _toList();  }
 
   static ClientReclaimAppInfoOverride decode(Object result) {
     result as List<Object?>;
@@ -477,12 +524,17 @@ class ClientReclaimAppInfoOverride {
 
   @override
   // ignore: avoid_equals_and_hash_code_on_mutable_classes
-  int get hashCode => Object.hashAll(_toList());
+  int get hashCode => Object.hashAll(_toList())
+;
 }
 
 /// Identification information of a session.
 class ReclaimSessionIdentityUpdate {
-  ReclaimSessionIdentityUpdate({required this.appId, required this.providerId, required this.sessionId});
+  ReclaimSessionIdentityUpdate({
+    required this.appId,
+    required this.providerId,
+    required this.sessionId,
+  });
 
   /// The application id.
   String appId;
@@ -494,12 +546,15 @@ class ReclaimSessionIdentityUpdate {
   String sessionId;
 
   List<Object?> _toList() {
-    return <Object?>[appId, providerId, sessionId];
+    return <Object?>[
+      appId,
+      providerId,
+      sessionId,
+    ];
   }
 
   Object encode() {
-    return _toList();
-  }
+    return _toList();  }
 
   static ReclaimSessionIdentityUpdate decode(Object result) {
     result as List<Object?>;
@@ -524,16 +579,17 @@ class ReclaimSessionIdentityUpdate {
 
   @override
   // ignore: avoid_equals_and_hash_code_on_mutable_classes
-  int get hashCode => Object.hashAll(_toList());
+  int get hashCode => Object.hashAll(_toList())
+;
 }
 
 class ReclaimApiVerificationOptions {
   ReclaimApiVerificationOptions({
-    required this.canDeleteCookiesBeforeVerificationStarts,
-    required this.canUseAttestorAuthenticationRequest,
-    required this.claimCreationType,
-    required this.canAutoSubmit,
-    required this.isCloseButtonVisible,
+    this.canDeleteCookiesBeforeVerificationStarts = true,
+    this.canUseAttestorAuthenticationRequest = false,
+    this.claimCreationType = ClaimCreationTypeApi.standalone,
+    this.canAutoSubmit = true,
+    this.isCloseButtonVisible = true,
   });
 
   /// Whether to delete cookies before user journey starts in the client web view.
@@ -566,8 +622,7 @@ class ReclaimApiVerificationOptions {
   }
 
   Object encode() {
-    return _toList();
-  }
+    return _toList();  }
 
   static ReclaimApiVerificationOptions decode(Object result) {
     result as List<Object?>;
@@ -594,8 +649,173 @@ class ReclaimApiVerificationOptions {
 
   @override
   // ignore: avoid_equals_and_hash_code_on_mutable_classes
-  int get hashCode => Object.hashAll(_toList());
+  int get hashCode => Object.hashAll(_toList())
+;
 }
+
+class ProviderVersionApi {
+  ProviderVersionApi({
+    this.versionExpression,
+    this.resolvedVersion,
+  });
+
+  String? versionExpression;
+
+  String? resolvedVersion;
+
+  List<Object?> _toList() {
+    return <Object?>[
+      versionExpression,
+      resolvedVersion,
+    ];
+  }
+
+  Object encode() {
+    return _toList();  }
+
+  static ProviderVersionApi decode(Object result) {
+    result as List<Object?>;
+    return ProviderVersionApi(
+      versionExpression: result[0] as String?,
+      resolvedVersion: result[1] as String?,
+    );
+  }
+
+  @override
+  // ignore: avoid_equals_and_hash_code_on_mutable_classes
+  bool operator ==(Object other) {
+    if (other is! ProviderVersionApi || other.runtimeType != runtimeType) {
+      return false;
+    }
+    if (identical(this, other)) {
+      return true;
+    }
+    return _deepEquals(encode(), other.encode());
+  }
+
+  @override
+  // ignore: avoid_equals_and_hash_code_on_mutable_classes
+  int get hashCode => Object.hashAll(_toList())
+;
+}
+
+class SessionInitResponseApi {
+  SessionInitResponseApi({
+    required this.sessionId,
+    this.resolvedProviderVersion,
+  });
+
+  String sessionId;
+
+  String? resolvedProviderVersion;
+
+  List<Object?> _toList() {
+    return <Object?>[
+      sessionId,
+      resolvedProviderVersion,
+    ];
+  }
+
+  Object encode() {
+    return _toList();  }
+
+  static SessionInitResponseApi decode(Object result) {
+    result as List<Object?>;
+    return SessionInitResponseApi(
+      sessionId: result[0]! as String,
+      resolvedProviderVersion: result[1] as String?,
+    );
+  }
+
+  @override
+  // ignore: avoid_equals_and_hash_code_on_mutable_classes
+  bool operator ==(Object other) {
+    if (other is! SessionInitResponseApi || other.runtimeType != runtimeType) {
+      return false;
+    }
+    if (identical(this, other)) {
+      return true;
+    }
+    return _deepEquals(encode(), other.encode());
+  }
+
+  @override
+  // ignore: avoid_equals_and_hash_code_on_mutable_classes
+  int get hashCode => Object.hashAll(_toList())
+;
+}
+
+class LogEntryApi {
+  LogEntryApi({
+    this.sessionId,
+    required this.message,
+    required this.level,
+    required this.dateTimeIso,
+    required this.source,
+    this.error,
+    this.stackTraceAsString,
+  });
+
+  String? sessionId;
+
+  String message;
+
+  int level;
+
+  String dateTimeIso;
+
+  String source;
+
+  String? error;
+
+  String? stackTraceAsString;
+
+  List<Object?> _toList() {
+    return <Object?>[
+      sessionId,
+      message,
+      level,
+      dateTimeIso,
+      source,
+      error,
+      stackTraceAsString,
+    ];
+  }
+
+  Object encode() {
+    return _toList();  }
+
+  static LogEntryApi decode(Object result) {
+    result as List<Object?>;
+    return LogEntryApi(
+      sessionId: result[0] as String?,
+      message: result[1]! as String,
+      level: result[2]! as int,
+      dateTimeIso: result[3]! as String,
+      source: result[4]! as String,
+      error: result[5] as String?,
+      stackTraceAsString: result[6] as String?,
+    );
+  }
+
+  @override
+  // ignore: avoid_equals_and_hash_code_on_mutable_classes
+  bool operator ==(Object other) {
+    if (other is! LogEntryApi || other.runtimeType != runtimeType) {
+      return false;
+    }
+    if (identical(this, other)) {
+      return true;
+    }
+    return _deepEquals(encode(), other.encode());
+  }
+
+  @override
+  // ignore: avoid_equals_and_hash_code_on_mutable_classes
+  int get hashCode => Object.hashAll(_toList())
+;
+}
+
 
 class _PigeonCodec extends StandardMessageCodec {
   const _PigeonCodec();
@@ -604,44 +824,53 @@ class _PigeonCodec extends StandardMessageCodec {
     if (value is int) {
       buffer.putUint8(4);
       buffer.putInt64(value);
-    } else if (value is ReclaimApiVerificationExceptionType) {
+    }    else if (value is ReclaimApiVerificationExceptionType) {
       buffer.putUint8(129);
       writeValue(buffer, value.index);
-    } else if (value is ReclaimSessionStatus) {
+    }    else if (value is ReclaimSessionStatus) {
       buffer.putUint8(130);
       writeValue(buffer, value.index);
-    } else if (value is ClaimCreationTypeApi) {
+    }    else if (value is ClaimCreationTypeApi) {
       buffer.putUint8(131);
       writeValue(buffer, value.index);
-    } else if (value is ReclaimApiVerificationRequest) {
+    }    else if (value is ReclaimApiVerificationRequest) {
       buffer.putUint8(132);
       writeValue(buffer, value.encode());
-    } else if (value is ReclaimApiVerificationException) {
+    }    else if (value is ReclaimApiVerificationException) {
       buffer.putUint8(133);
       writeValue(buffer, value.encode());
-    } else if (value is ReclaimApiVerificationResponse) {
+    }    else if (value is ReclaimApiVerificationResponse) {
       buffer.putUint8(134);
       writeValue(buffer, value.encode());
-    } else if (value is ClientProviderInformationOverride) {
+    }    else if (value is ClientProviderInformationOverride) {
       buffer.putUint8(135);
       writeValue(buffer, value.encode());
-    } else if (value is ClientFeatureOverrides) {
+    }    else if (value is ClientFeatureOverrides) {
       buffer.putUint8(136);
       writeValue(buffer, value.encode());
-    } else if (value is ClientLogConsumerOverride) {
+    }    else if (value is ClientLogConsumerOverride) {
       buffer.putUint8(137);
       writeValue(buffer, value.encode());
-    } else if (value is ClientReclaimSessionManagementOverride) {
+    }    else if (value is ClientReclaimSessionManagementOverride) {
       buffer.putUint8(138);
       writeValue(buffer, value.encode());
-    } else if (value is ClientReclaimAppInfoOverride) {
+    }    else if (value is ClientReclaimAppInfoOverride) {
       buffer.putUint8(139);
       writeValue(buffer, value.encode());
-    } else if (value is ReclaimSessionIdentityUpdate) {
+    }    else if (value is ReclaimSessionIdentityUpdate) {
       buffer.putUint8(140);
       writeValue(buffer, value.encode());
-    } else if (value is ReclaimApiVerificationOptions) {
+    }    else if (value is ReclaimApiVerificationOptions) {
       buffer.putUint8(141);
+      writeValue(buffer, value.encode());
+    }    else if (value is ProviderVersionApi) {
+      buffer.putUint8(142);
+      writeValue(buffer, value.encode());
+    }    else if (value is SessionInitResponseApi) {
+      buffer.putUint8(143);
+      writeValue(buffer, value.encode());
+    }    else if (value is LogEntryApi) {
+      buffer.putUint8(144);
       writeValue(buffer, value.encode());
     } else {
       super.writeValue(buffer, value);
@@ -651,35 +880,41 @@ class _PigeonCodec extends StandardMessageCodec {
   @override
   Object? readValueOfType(int type, ReadBuffer buffer) {
     switch (type) {
-      case 129:
+      case 129: 
         final int? value = readValue(buffer) as int?;
         return value == null ? null : ReclaimApiVerificationExceptionType.values[value];
-      case 130:
+      case 130: 
         final int? value = readValue(buffer) as int?;
         return value == null ? null : ReclaimSessionStatus.values[value];
-      case 131:
+      case 131: 
         final int? value = readValue(buffer) as int?;
         return value == null ? null : ClaimCreationTypeApi.values[value];
-      case 132:
+      case 132: 
         return ReclaimApiVerificationRequest.decode(readValue(buffer)!);
-      case 133:
+      case 133: 
         return ReclaimApiVerificationException.decode(readValue(buffer)!);
-      case 134:
+      case 134: 
         return ReclaimApiVerificationResponse.decode(readValue(buffer)!);
-      case 135:
+      case 135: 
         return ClientProviderInformationOverride.decode(readValue(buffer)!);
-      case 136:
+      case 136: 
         return ClientFeatureOverrides.decode(readValue(buffer)!);
-      case 137:
+      case 137: 
         return ClientLogConsumerOverride.decode(readValue(buffer)!);
-      case 138:
+      case 138: 
         return ClientReclaimSessionManagementOverride.decode(readValue(buffer)!);
-      case 139:
+      case 139: 
         return ClientReclaimAppInfoOverride.decode(readValue(buffer)!);
-      case 140:
+      case 140: 
         return ReclaimSessionIdentityUpdate.decode(readValue(buffer)!);
-      case 141:
+      case 141: 
         return ReclaimApiVerificationOptions.decode(readValue(buffer)!);
+      case 142: 
+        return ProviderVersionApi.decode(readValue(buffer)!);
+      case 143: 
+        return SessionInitResponseApi.decode(readValue(buffer)!);
+      case 144: 
+        return LogEntryApi.decode(readValue(buffer)!);
       default:
         return super.readValueOfType(type, buffer);
     }
@@ -694,49 +929,40 @@ abstract class ReclaimModuleApi {
 
   Future<ReclaimApiVerificationResponse> startVerificationFromUrl(String url);
 
-  Future<void> setOverrides(
-    ClientProviderInformationOverride? provider,
-    ClientFeatureOverrides? feature,
-    ClientLogConsumerOverride? logConsumer,
-    ClientReclaimSessionManagementOverride? sessionManagement,
-    ClientReclaimAppInfoOverride? appInfo,
-    String? capabilityAccessToken,
-  );
+  Future<ReclaimApiVerificationResponse> startVerificationFromJson(Map<dynamic, dynamic> template);
+
+  Future<void> setOverrides(ClientProviderInformationOverride? provider, ClientFeatureOverrides? feature, ClientLogConsumerOverride? logConsumer, ClientReclaimSessionManagementOverride? sessionManagement, ClientReclaimAppInfoOverride? appInfo, String? capabilityAccessToken);
 
   Future<void> clearAllOverrides();
 
   Future<void> setVerificationOptions(ReclaimApiVerificationOptions? options);
 
+  Future<bool> sendLog(LogEntryApi entry);
+
   Future<bool> ping();
 
-  static void setUp(ReclaimModuleApi? api, {BinaryMessenger? binaryMessenger, String messageChannelSuffix = ''}) {
+  static void setUp(ReclaimModuleApi? api, {BinaryMessenger? binaryMessenger, String messageChannelSuffix = '',}) {
     messageChannelSuffix = messageChannelSuffix.isNotEmpty ? '.$messageChannelSuffix' : '';
     {
       final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.startVerification$messageChannelSuffix',
-        pigeonChannelCodec,
-        binaryMessenger: binaryMessenger,
-      );
+          'dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.startVerification$messageChannelSuffix', pigeonChannelCodec,
+          binaryMessenger: binaryMessenger);
       if (api == null) {
         pigeonVar_channel.setMessageHandler(null);
       } else {
         pigeonVar_channel.setMessageHandler((Object? message) async {
-          assert(
-            message != null,
-            'Argument for dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.startVerification was null.',
-          );
+          assert(message != null,
+          'Argument for dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.startVerification was null.');
           final List<Object?> args = (message as List<Object?>?)!;
           final ReclaimApiVerificationRequest? arg_request = (args[0] as ReclaimApiVerificationRequest?);
-          assert(
-            arg_request != null,
-            'Argument for dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.startVerification was null, expected non-null ReclaimApiVerificationRequest.',
-          );
+          assert(arg_request != null,
+              'Argument for dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.startVerification was null, expected non-null ReclaimApiVerificationRequest.');
           try {
             final ReclaimApiVerificationResponse output = await api.startVerification(arg_request!);
             return wrapResponse(result: output);
           } on PlatformException catch (e) {
             return wrapResponse(error: e);
-          } catch (e) {
+          }          catch (e) {
             return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
           }
         });
@@ -744,30 +970,24 @@ abstract class ReclaimModuleApi {
     }
     {
       final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.startVerificationFromUrl$messageChannelSuffix',
-        pigeonChannelCodec,
-        binaryMessenger: binaryMessenger,
-      );
+          'dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.startVerificationFromUrl$messageChannelSuffix', pigeonChannelCodec,
+          binaryMessenger: binaryMessenger);
       if (api == null) {
         pigeonVar_channel.setMessageHandler(null);
       } else {
         pigeonVar_channel.setMessageHandler((Object? message) async {
-          assert(
-            message != null,
-            'Argument for dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.startVerificationFromUrl was null.',
-          );
+          assert(message != null,
+          'Argument for dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.startVerificationFromUrl was null.');
           final List<Object?> args = (message as List<Object?>?)!;
           final String? arg_url = (args[0] as String?);
-          assert(
-            arg_url != null,
-            'Argument for dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.startVerificationFromUrl was null, expected non-null String.',
-          );
+          assert(arg_url != null,
+              'Argument for dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.startVerificationFromUrl was null, expected non-null String.');
           try {
             final ReclaimApiVerificationResponse output = await api.startVerificationFromUrl(arg_url!);
             return wrapResponse(result: output);
           } on PlatformException catch (e) {
             return wrapResponse(error: e);
-          } catch (e) {
+          }          catch (e) {
             return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
           }
         });
@@ -775,39 +995,24 @@ abstract class ReclaimModuleApi {
     }
     {
       final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.setOverrides$messageChannelSuffix',
-        pigeonChannelCodec,
-        binaryMessenger: binaryMessenger,
-      );
+          'dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.startVerificationFromJson$messageChannelSuffix', pigeonChannelCodec,
+          binaryMessenger: binaryMessenger);
       if (api == null) {
         pigeonVar_channel.setMessageHandler(null);
       } else {
         pigeonVar_channel.setMessageHandler((Object? message) async {
-          assert(
-            message != null,
-            'Argument for dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.setOverrides was null.',
-          );
+          assert(message != null,
+          'Argument for dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.startVerificationFromJson was null.');
           final List<Object?> args = (message as List<Object?>?)!;
-          final ClientProviderInformationOverride? arg_provider = (args[0] as ClientProviderInformationOverride?);
-          final ClientFeatureOverrides? arg_feature = (args[1] as ClientFeatureOverrides?);
-          final ClientLogConsumerOverride? arg_logConsumer = (args[2] as ClientLogConsumerOverride?);
-          final ClientReclaimSessionManagementOverride? arg_sessionManagement =
-              (args[3] as ClientReclaimSessionManagementOverride?);
-          final ClientReclaimAppInfoOverride? arg_appInfo = (args[4] as ClientReclaimAppInfoOverride?);
-          final String? arg_capabilityAccessToken = (args[5] as String?);
+          final Map<dynamic, dynamic>? arg_template = (args[0] as Map<Object?, Object?>?)?.cast<dynamic, dynamic>();
+          assert(arg_template != null,
+              'Argument for dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.startVerificationFromJson was null, expected non-null Map<dynamic, dynamic>.');
           try {
-            await api.setOverrides(
-              arg_provider,
-              arg_feature,
-              arg_logConsumer,
-              arg_sessionManagement,
-              arg_appInfo,
-              arg_capabilityAccessToken,
-            );
-            return wrapResponse(empty: true);
+            final ReclaimApiVerificationResponse output = await api.startVerificationFromJson(arg_template!);
+            return wrapResponse(result: output);
           } on PlatformException catch (e) {
             return wrapResponse(error: e);
-          } catch (e) {
+          }          catch (e) {
             return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
           }
         });
@@ -815,10 +1020,36 @@ abstract class ReclaimModuleApi {
     }
     {
       final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.clearAllOverrides$messageChannelSuffix',
-        pigeonChannelCodec,
-        binaryMessenger: binaryMessenger,
-      );
+          'dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.setOverrides$messageChannelSuffix', pigeonChannelCodec,
+          binaryMessenger: binaryMessenger);
+      if (api == null) {
+        pigeonVar_channel.setMessageHandler(null);
+      } else {
+        pigeonVar_channel.setMessageHandler((Object? message) async {
+          assert(message != null,
+          'Argument for dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.setOverrides was null.');
+          final List<Object?> args = (message as List<Object?>?)!;
+          final ClientProviderInformationOverride? arg_provider = (args[0] as ClientProviderInformationOverride?);
+          final ClientFeatureOverrides? arg_feature = (args[1] as ClientFeatureOverrides?);
+          final ClientLogConsumerOverride? arg_logConsumer = (args[2] as ClientLogConsumerOverride?);
+          final ClientReclaimSessionManagementOverride? arg_sessionManagement = (args[3] as ClientReclaimSessionManagementOverride?);
+          final ClientReclaimAppInfoOverride? arg_appInfo = (args[4] as ClientReclaimAppInfoOverride?);
+          final String? arg_capabilityAccessToken = (args[5] as String?);
+          try {
+            await api.setOverrides(arg_provider, arg_feature, arg_logConsumer, arg_sessionManagement, arg_appInfo, arg_capabilityAccessToken);
+            return wrapResponse(empty: true);
+          } on PlatformException catch (e) {
+            return wrapResponse(error: e);
+          }          catch (e) {
+            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          }
+        });
+      }
+    }
+    {
+      final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.clearAllOverrides$messageChannelSuffix', pigeonChannelCodec,
+          binaryMessenger: binaryMessenger);
       if (api == null) {
         pigeonVar_channel.setMessageHandler(null);
       } else {
@@ -828,7 +1059,7 @@ abstract class ReclaimModuleApi {
             return wrapResponse(empty: true);
           } on PlatformException catch (e) {
             return wrapResponse(error: e);
-          } catch (e) {
+          }          catch (e) {
             return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
           }
         });
@@ -836,18 +1067,14 @@ abstract class ReclaimModuleApi {
     }
     {
       final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.setVerificationOptions$messageChannelSuffix',
-        pigeonChannelCodec,
-        binaryMessenger: binaryMessenger,
-      );
+          'dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.setVerificationOptions$messageChannelSuffix', pigeonChannelCodec,
+          binaryMessenger: binaryMessenger);
       if (api == null) {
         pigeonVar_channel.setMessageHandler(null);
       } else {
         pigeonVar_channel.setMessageHandler((Object? message) async {
-          assert(
-            message != null,
-            'Argument for dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.setVerificationOptions was null.',
-          );
+          assert(message != null,
+          'Argument for dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.setVerificationOptions was null.');
           final List<Object?> args = (message as List<Object?>?)!;
           final ReclaimApiVerificationOptions? arg_options = (args[0] as ReclaimApiVerificationOptions?);
           try {
@@ -855,7 +1082,7 @@ abstract class ReclaimModuleApi {
             return wrapResponse(empty: true);
           } on PlatformException catch (e) {
             return wrapResponse(error: e);
-          } catch (e) {
+          }          catch (e) {
             return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
           }
         });
@@ -863,10 +1090,33 @@ abstract class ReclaimModuleApi {
     }
     {
       final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.ping$messageChannelSuffix',
-        pigeonChannelCodec,
-        binaryMessenger: binaryMessenger,
-      );
+          'dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.sendLog$messageChannelSuffix', pigeonChannelCodec,
+          binaryMessenger: binaryMessenger);
+      if (api == null) {
+        pigeonVar_channel.setMessageHandler(null);
+      } else {
+        pigeonVar_channel.setMessageHandler((Object? message) async {
+          assert(message != null,
+          'Argument for dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.sendLog was null.');
+          final List<Object?> args = (message as List<Object?>?)!;
+          final LogEntryApi? arg_entry = (args[0] as LogEntryApi?);
+          assert(arg_entry != null,
+              'Argument for dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.sendLog was null, expected non-null LogEntryApi.');
+          try {
+            final bool output = await api.sendLog(arg_entry!);
+            return wrapResponse(result: output);
+          } on PlatformException catch (e) {
+            return wrapResponse(error: e);
+          }          catch (e) {
+            return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
+          }
+        });
+      }
+    }
+    {
+      final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
+          'dev.flutter.pigeon.reclaim_verifier_module.ReclaimModuleApi.ping$messageChannelSuffix', pigeonChannelCodec,
+          binaryMessenger: binaryMessenger);
       if (api == null) {
         pigeonVar_channel.setMessageHandler(null);
       } else {
@@ -876,7 +1126,7 @@ abstract class ReclaimModuleApi {
             return wrapResponse(result: output);
           } on PlatformException catch (e) {
             return wrapResponse(error: e);
-          } catch (e) {
+          }          catch (e) {
             return wrapResponse(error: PlatformException(code: 'error', message: e.toString()));
           }
         });
@@ -891,8 +1141,8 @@ class ReclaimHostOverridesApi {
   /// available for dependency injection.  If it is left null, the default
   /// BinaryMessenger will be used which routes to the host platform.
   ReclaimHostOverridesApi({BinaryMessenger? binaryMessenger, String messageChannelSuffix = ''})
-    : pigeonVar_binaryMessenger = binaryMessenger,
-      pigeonVar_messageChannelSuffix = messageChannelSuffix.isNotEmpty ? '.$messageChannelSuffix' : '';
+      : pigeonVar_binaryMessenger = binaryMessenger,
+        pigeonVar_messageChannelSuffix = messageChannelSuffix.isNotEmpty ? '.$messageChannelSuffix' : '';
   final BinaryMessenger? pigeonVar_binaryMessenger;
 
   static const MessageCodec<Object?> pigeonChannelCodec = _PigeonCodec();
@@ -900,15 +1150,15 @@ class ReclaimHostOverridesApi {
   final String pigeonVar_messageChannelSuffix;
 
   Future<void> onLogs(String logJsonString) async {
-    final String pigeonVar_channelName =
-        'dev.flutter.pigeon.reclaim_verifier_module.ReclaimHostOverridesApi.onLogs$pigeonVar_messageChannelSuffix';
+    final String pigeonVar_channelName = 'dev.flutter.pigeon.reclaim_verifier_module.ReclaimHostOverridesApi.onLogs$pigeonVar_messageChannelSuffix';
     final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
       pigeonVar_channelName,
       pigeonChannelCodec,
       binaryMessenger: pigeonVar_binaryMessenger,
     );
     final Future<Object?> pigeonVar_sendFuture = pigeonVar_channel.send(<Object?>[logJsonString]);
-    final List<Object?>? pigeonVar_replyList = await pigeonVar_sendFuture as List<Object?>?;
+    final List<Object?>? pigeonVar_replyList =
+        await pigeonVar_sendFuture as List<Object?>?;
     if (pigeonVar_replyList == null) {
       throw _createConnectionError(pigeonVar_channelName);
     } else if (pigeonVar_replyList.length > 1) {
@@ -922,26 +1172,16 @@ class ReclaimHostOverridesApi {
     }
   }
 
-  Future<String> createSession({
-    required String appId,
-    required String providerId,
-    required String timestamp,
-    required String signature,
-  }) async {
-    final String pigeonVar_channelName =
-        'dev.flutter.pigeon.reclaim_verifier_module.ReclaimHostOverridesApi.createSession$pigeonVar_messageChannelSuffix';
+  Future<SessionInitResponseApi> createSession({required String appId, required String providerId, required String timestamp, required String signature, required String providerVersion, }) async {
+    final String pigeonVar_channelName = 'dev.flutter.pigeon.reclaim_verifier_module.ReclaimHostOverridesApi.createSession$pigeonVar_messageChannelSuffix';
     final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
       pigeonVar_channelName,
       pigeonChannelCodec,
       binaryMessenger: pigeonVar_binaryMessenger,
     );
-    final Future<Object?> pigeonVar_sendFuture = pigeonVar_channel.send(<Object?>[
-      appId,
-      providerId,
-      timestamp,
-      signature,
-    ]);
-    final List<Object?>? pigeonVar_replyList = await pigeonVar_sendFuture as List<Object?>?;
+    final Future<Object?> pigeonVar_sendFuture = pigeonVar_channel.send(<Object?>[appId, providerId, timestamp, signature, providerVersion]);
+    final List<Object?>? pigeonVar_replyList =
+        await pigeonVar_sendFuture as List<Object?>?;
     if (pigeonVar_replyList == null) {
       throw _createConnectionError(pigeonVar_channelName);
     } else if (pigeonVar_replyList.length > 1) {
@@ -956,20 +1196,20 @@ class ReclaimHostOverridesApi {
         message: 'Host platform returned null value for non-null return value.',
       );
     } else {
-      return (pigeonVar_replyList[0] as String?)!;
+      return (pigeonVar_replyList[0] as SessionInitResponseApi?)!;
     }
   }
 
   Future<bool> updateSession({required String sessionId, required ReclaimSessionStatus status}) async {
-    final String pigeonVar_channelName =
-        'dev.flutter.pigeon.reclaim_verifier_module.ReclaimHostOverridesApi.updateSession$pigeonVar_messageChannelSuffix';
+    final String pigeonVar_channelName = 'dev.flutter.pigeon.reclaim_verifier_module.ReclaimHostOverridesApi.updateSession$pigeonVar_messageChannelSuffix';
     final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
       pigeonVar_channelName,
       pigeonChannelCodec,
       binaryMessenger: pigeonVar_binaryMessenger,
     );
     final Future<Object?> pigeonVar_sendFuture = pigeonVar_channel.send(<Object?>[sessionId, status]);
-    final List<Object?>? pigeonVar_replyList = await pigeonVar_sendFuture as List<Object?>?;
+    final List<Object?>? pigeonVar_replyList =
+        await pigeonVar_sendFuture as List<Object?>?;
     if (pigeonVar_replyList == null) {
       throw _createConnectionError(pigeonVar_channelName);
     } else if (pigeonVar_replyList.length > 1) {
@@ -988,28 +1228,16 @@ class ReclaimHostOverridesApi {
     }
   }
 
-  Future<void> logSession({
-    required String appId,
-    required String providerId,
-    required String sessionId,
-    required String logType,
-    Map<String, dynamic>? metadata,
-  }) async {
-    final String pigeonVar_channelName =
-        'dev.flutter.pigeon.reclaim_verifier_module.ReclaimHostOverridesApi.logSession$pigeonVar_messageChannelSuffix';
+  Future<void> logSession({required String appId, required String providerId, required String sessionId, required String logType, Map<String, dynamic>? metadata, }) async {
+    final String pigeonVar_channelName = 'dev.flutter.pigeon.reclaim_verifier_module.ReclaimHostOverridesApi.logSession$pigeonVar_messageChannelSuffix';
     final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
       pigeonVar_channelName,
       pigeonChannelCodec,
       binaryMessenger: pigeonVar_binaryMessenger,
     );
-    final Future<Object?> pigeonVar_sendFuture = pigeonVar_channel.send(<Object?>[
-      appId,
-      providerId,
-      sessionId,
-      logType,
-      metadata,
-    ]);
-    final List<Object?>? pigeonVar_replyList = await pigeonVar_sendFuture as List<Object?>?;
+    final Future<Object?> pigeonVar_sendFuture = pigeonVar_channel.send(<Object?>[appId, providerId, sessionId, logType, metadata]);
+    final List<Object?>? pigeonVar_replyList =
+        await pigeonVar_sendFuture as List<Object?>?;
     if (pigeonVar_replyList == null) {
       throw _createConnectionError(pigeonVar_channelName);
     } else if (pigeonVar_replyList.length > 1) {
@@ -1024,15 +1252,15 @@ class ReclaimHostOverridesApi {
   }
 
   Future<void> onSessionIdentityUpdate(ReclaimSessionIdentityUpdate? update) async {
-    final String pigeonVar_channelName =
-        'dev.flutter.pigeon.reclaim_verifier_module.ReclaimHostOverridesApi.onSessionIdentityUpdate$pigeonVar_messageChannelSuffix';
+    final String pigeonVar_channelName = 'dev.flutter.pigeon.reclaim_verifier_module.ReclaimHostOverridesApi.onSessionIdentityUpdate$pigeonVar_messageChannelSuffix';
     final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
       pigeonVar_channelName,
       pigeonChannelCodec,
       binaryMessenger: pigeonVar_binaryMessenger,
     );
     final Future<Object?> pigeonVar_sendFuture = pigeonVar_channel.send(<Object?>[update]);
-    final List<Object?>? pigeonVar_replyList = await pigeonVar_sendFuture as List<Object?>?;
+    final List<Object?>? pigeonVar_replyList =
+        await pigeonVar_sendFuture as List<Object?>?;
     if (pigeonVar_replyList == null) {
       throw _createConnectionError(pigeonVar_channelName);
     } else if (pigeonVar_replyList.length > 1) {
@@ -1046,28 +1274,16 @@ class ReclaimHostOverridesApi {
     }
   }
 
-  Future<String> fetchProviderInformation({
-    required String appId,
-    required String providerId,
-    required String sessionId,
-    required String signature,
-    required String timestamp,
-  }) async {
-    final String pigeonVar_channelName =
-        'dev.flutter.pigeon.reclaim_verifier_module.ReclaimHostOverridesApi.fetchProviderInformation$pigeonVar_messageChannelSuffix';
+  Future<String> fetchProviderInformation({required String appId, required String providerId, required String sessionId, required String signature, required String timestamp, }) async {
+    final String pigeonVar_channelName = 'dev.flutter.pigeon.reclaim_verifier_module.ReclaimHostOverridesApi.fetchProviderInformation$pigeonVar_messageChannelSuffix';
     final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
       pigeonVar_channelName,
       pigeonChannelCodec,
       binaryMessenger: pigeonVar_binaryMessenger,
     );
-    final Future<Object?> pigeonVar_sendFuture = pigeonVar_channel.send(<Object?>[
-      appId,
-      providerId,
-      sessionId,
-      signature,
-      timestamp,
-    ]);
-    final List<Object?>? pigeonVar_replyList = await pigeonVar_sendFuture as List<Object?>?;
+    final Future<Object?> pigeonVar_sendFuture = pigeonVar_channel.send(<Object?>[appId, providerId, sessionId, signature, timestamp]);
+    final List<Object?>? pigeonVar_replyList =
+        await pigeonVar_sendFuture as List<Object?>?;
     if (pigeonVar_replyList == null) {
       throw _createConnectionError(pigeonVar_channelName);
     } else if (pigeonVar_replyList.length > 1) {
@@ -1092,8 +1308,8 @@ class ReclaimHostVerificationApi {
   /// available for dependency injection.  If it is left null, the default
   /// BinaryMessenger will be used which routes to the host platform.
   ReclaimHostVerificationApi({BinaryMessenger? binaryMessenger, String messageChannelSuffix = ''})
-    : pigeonVar_binaryMessenger = binaryMessenger,
-      pigeonVar_messageChannelSuffix = messageChannelSuffix.isNotEmpty ? '.$messageChannelSuffix' : '';
+      : pigeonVar_binaryMessenger = binaryMessenger,
+        pigeonVar_messageChannelSuffix = messageChannelSuffix.isNotEmpty ? '.$messageChannelSuffix' : '';
   final BinaryMessenger? pigeonVar_binaryMessenger;
 
   static const MessageCodec<Object?> pigeonChannelCodec = _PigeonCodec();
@@ -1101,15 +1317,15 @@ class ReclaimHostVerificationApi {
   final String pigeonVar_messageChannelSuffix;
 
   Future<String> fetchAttestorAuthenticationRequest(Map<dynamic, dynamic> reclaimHttpProvider) async {
-    final String pigeonVar_channelName =
-        'dev.flutter.pigeon.reclaim_verifier_module.ReclaimHostVerificationApi.fetchAttestorAuthenticationRequest$pigeonVar_messageChannelSuffix';
+    final String pigeonVar_channelName = 'dev.flutter.pigeon.reclaim_verifier_module.ReclaimHostVerificationApi.fetchAttestorAuthenticationRequest$pigeonVar_messageChannelSuffix';
     final BasicMessageChannel<Object?> pigeonVar_channel = BasicMessageChannel<Object?>(
       pigeonVar_channelName,
       pigeonChannelCodec,
       binaryMessenger: pigeonVar_binaryMessenger,
     );
     final Future<Object?> pigeonVar_sendFuture = pigeonVar_channel.send(<Object?>[reclaimHttpProvider]);
-    final List<Object?>? pigeonVar_replyList = await pigeonVar_sendFuture as List<Object?>?;
+    final List<Object?>? pigeonVar_replyList =
+        await pigeonVar_sendFuture as List<Object?>?;
     if (pigeonVar_replyList == null) {
       throw _createConnectionError(pigeonVar_channelName);
     } else if (pigeonVar_replyList.length > 1) {